Refactor : Restructure DTO packages for board and task into separate subpackages.
