+++ src/main/java/org/trello/trelloclone/controllers/BoardController.java
@@ -5,7 +5,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.trello.trelloclone.dtos.common.ResponseObjectJsonDto;
-import org.trello.trelloclone.dtos.models.BoardRequestDto;
+import org.trello.trelloclone.dtos.models.board.BoardRequestDto;
 import org.trello.trelloclone.service.BoardService;
 
 @RestController
