+++ src/main/java/org/trello/trelloclone/controllers/TaskController.java
@@ -5,8 +5,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.trello.trelloclone.dtos.common.ResponseObjectJsonDto;
-import org.trello.trelloclone.dtos.models.TaskRequestDto;
-import org.trello.trelloclone.models.Task;
+import org.trello.trelloclone.dtos.models.task.TaskRequestDto;
 import org.trello.trelloclone.service.TaskService;
 
 @RestController
