+++ src/main/java/org/trello/trelloclone/dtos/models/board/BoardResponseDto.java
@@ -1,4 +1,4 @@
-package org.trello.trelloclone.dtos.models;
+package org.trello.trelloclone.dtos.models.board;
 
 import lombok.Data;
 import java.util.List;
