+++ src/main/java/org/trello/trelloclone/dtos/models/task/TaskRequestDto.java
@@ -1,11 +1,10 @@
-package org.trello.trelloclone.dtos.models;
+package org.trello.trelloclone.dtos.models.task;
 
 import lombok.Data;
 import java.time.LocalDate;
 
 @Data
-public class TaskResponseDto {
-    private Long id;
+public class TaskRequestDto {
     private String title;
     private String description;
     private String status;
