+++ src/main/java/org/trello/trelloclone/dtos/models/task/TaskResponseDto.java
@@ -1,10 +1,11 @@
-package org.trello.trelloclone.dtos.models;
+package org.trello.trelloclone.dtos.models.task;
 
 import lombok.Data;
 import java.time.LocalDate;
 
 @Data
-public class TaskRequestDto {
+public class TaskResponseDto {
+    private Long id;
     private String title;
     private String description;
     private String status;
